backend/requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
aiomysql==0.2.0
SQLAlchemy==2.0.36
alembic==1.12.1
PyMySQL==1.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
python-dotenv==1.0.0

# WebSocket
websockets==12.0

# Email
aiosmtplib==3.0.1

# File Upload
aiofiles==23.2.1
# pillow==11.0.0  # Requires Python 3.13 or zlib dev tools on Windows

# Validation
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Caching
redis==5.0.1
hiredis==2.2.3

# Utilities
python-dateutil==2.8.2
httpx==0.25.2
jinja2==3.1.2

# PDF Generation
reportlab==4.0.7
weasyprint==60.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
